Memory footprint reduced. Matched primer statistical data now included.

+++ src/main/java/CSV.java
@@ -19,6 +19,7 @@ import java.util.*;
 
  * Created by Charles Gregory on 12/8/2015. Controls .csv file writing and reading.
  */
+@SuppressWarnings("Duplicates")
 public class CSV {
     private static final String COMMA_DELIMITER = ",";
     //creates a directory
@@ -28,7 +29,7 @@ public class CSV {
         }
     }
     //Writes Common csv data
-    public static void writeCommonCSV(String s,Set<CharSequence> data) throws IOException {
+    public static void writeCommonCSV(String s, Set<String> data) throws IOException {
         String base = new File("").getAbsolutePath();
         FileWriter fileWriter = new FileWriter(base+"\\Common\\"+s+".csv");
         data.forEach(x ->{
@@ -45,17 +46,20 @@ public class CSV {
     //Writes phageData csv data
     public static void writeDataCSV(String s,Set<CharSequence> data,int bps) throws IOException {
         String base = new File("").getAbsolutePath();
-        FileWriter fileWriter = new FileWriter(base+"/PhageData/"+Integer.toString(bps)+s+".csv");
-        data.forEach(x ->{
-            try {
-                fileWriter.append(x);
-                fileWriter.append(COMMA_DELIMITER);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        });
-        fileWriter.flush();
-        fileWriter.close();
+        File file =new File(base+"/PhageData/"+Integer.toString(bps)+s+".csv");
+        if(!file.exists()) {
+            FileWriter fileWriter = new FileWriter(base + "/PhageData/" + Integer.toString(bps) + s + ".csv");
+            data.forEach(x -> {
+                try {
+                    fileWriter.append(x);
+                    fileWriter.append(COMMA_DELIMITER);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            });
+            fileWriter.flush();
+            fileWriter.close();
+        }
     }
     public static void writeDataCSV(String s,List<CharSequence> data,int bps) throws IOException {
         String base = new File("").getAbsolutePath();
@@ -72,7 +76,7 @@ public class CSV {
         fileWriter.close();
     }
     //Writes unique csv data
-    public static void writeUniqueCSV(String s,Set<CharSequence> data) throws IOException {
+    public static void writeUniqueCSV(String s, Set<String> data) throws IOException {
         String base = new File("").getAbsolutePath();
         FileWriter fileWriter = new FileWriter(base+"\\Unique\\"+s+".csv");
         data.forEach(x ->{
@@ -117,11 +121,11 @@ public class CSV {
         fileWriter.close();
     }
     //reads a csv written in any of the prior formats
-    public static Set<CharSequence> readCSV(String path1){
+    public static Set<String> readCSV(String path1){
         BufferedReader br = null;
         String line;
         String cvsSplitBy = ",";
-        CharSequence[] prim =null;
+        String[] prim =null;
         try {
             br = new BufferedReader( new FileReader(path1));
             while ((line = br.readLine()) != null) {
@@ -142,7 +146,7 @@ public class CSV {
                 }
             }
         }
-        Set<CharSequence> s = new HashSet<>();
+        Set<String> s = new HashSet<>();
         if(prim != null){
             Collections.addAll(s, prim);
         }
